[FIX]: optimized the code

+++ src/services/file.service.ts
@@ -1,7 +1,8 @@
 import db from '../database/index'
 import fs from 'fs'
+import { IFile } from "../interfaces/database.interface";
 
-async function getAll(id: number) {
+async function getAll(id: number) : Promise<IFile[]> {
   try {
     return db.File.findAll({ where: { user_id: id }})
   } catch(error) {
@@ -9,7 +10,7 @@ async function getAll(id: number) {
   }
 }
 
-async function save(id: number, file: any) {
+async function save(id: number, file: any) : Promise<void> {
   try {
     await db.File.create({
       user_id: id,
@@ -21,13 +22,15 @@ async function save(id: number, file: any) {
   }
 }
 
-async function destroy(id: number) {
+async function destroy(id: number) : Promise<void> {
   try {
     const file = await db.File.findByPk(id)
-    fs.unlinkSync(file.path)
-    await file.destroy({
-      force: true
-    })
+    if (file) {
+      fs.unlinkSync(file.path)
+      await file.destroy({
+        force: true
+      })
+    }
   } catch(error) {
     throw error
   }
